// src/sora/types.ts
export enum Operation {
  Swap = 'Swap',
  AddLiquidity = 'AddLiquidity',
  RemoveLiquidity = 'RemoveLiquidity',
  CreatePair = 'CreatePair',
  BatchAll = 'BatchAll',
  Faucet = 'Faucet',
  EthBridgeOutgoing = 'EthBridgeOutgoing',
  EthBridgeIncoming = 'EthBridgeIncoming',
  EvmOutgoing = 'EvmOutgoing',
  EvmIncoming = 'EvmIncoming',
  SubstrateOutgoing = 'SubstrateOutgoing',
  SubstrateIncoming = 'SubstrateIncoming',
  ClaimRewards = 'ClaimRewards',
  /** it's used for calc network fee */
  ClaimVestedRewards = 'ClaimVestedRewards',
  /** it's used for calc network fee */
  ClaimCrowdloanRewards = 'ClaimCrowdloanRewards',
  /** it's used for calc network fee */
  ClaimLiquidityProvisionRewards = 'LiquidityProvisionRewards',
  /** it's used for calc network fee */
  ClaimExternalRewards = 'ClaimExternalRewards',
  /** it's used for internal needs as the MST batch with transfers  */
  TransferAll = 'TransferAll',
  /** Complex Swap */
  SwapAndSend = 'SwapAndSend',
  SwapTransferBatch = 'SwapTransferBatch',
  /** Referral System */
  ReferralReserveXor = 'ReferralReserveXor',
  ReferralUnreserveXor = 'ReferralUnreserveXor',
  ReferralSetInvitedUser = 'ReferralSetInvitedUser',
  /** Staking */
  StakingBond = 'StakingBond',
  StakingBondAndNominate = 'StakingBondAndNominate',
  StakingBondExtra = 'StakingBondExtra',
  StakingRebond = 'StakingRebond',
  StakingUnbond = 'StakingUnbond',
  StakingWithdrawUnbonded = 'StakingWithdrawUnbonded',
  StakingNominate = 'StakingNominate',
  StakingChill = 'StakingChill',
  StakingSetPayee = 'StakingSetPayee',
  StakingSetController = 'StakingSetController',
  StakingPayout = 'StakingPayout',
  /** Demeter Farming Platform  */
  DemeterFarmingDepositLiquidity = 'DemeterFarmingDepositLiquidity',
  DemeterFarmingWithdrawLiquidity = 'DemeterFarmingWithdrawLiquidity',
  DemeterFarmingStakeToken = 'DemeterFarmingStakeToken',
  DemeterFarmingUnstakeToken = 'DemeterFarmingUnstakeToken',
  DemeterFarmingGetRewards = 'DemeterFarmingGetRewards',
  /** Ceres Liquidity Locker  */
  CeresLiquidityLockerLockLiquidity = 'CeresLiquidityLockerLockLiquidity',
  /** Order Book */
  OrderBookPlaceLimitOrder = 'OrderBookPlaceLimitOrder',
  OrderBookCancelLimitOrder = 'OrderBookCancelLimitOrder',
  OrderBookCancelLimitOrders = 'OrderBookCancelLimitOrders',
  /** Asset management */
  RegisterAsset = 'RegisterAsset',
  Transfer = 'Transfer',
  XorlessTransfer = 'XorlessTransfer',
  Mint = 'Mint',
  Burn = 'Burn',
  /** Kensetsu */
  CreateVault = 'CreateVault',
  CloseVault = 'CloseVault',
  RepayVaultDebt = 'RepayVaultDebt',
  DepositCollateral = 'DepositCollateral',
  BorrowVaultDebt = 'BorrowVaultDebt',
  /** DEFI-R */
  SetAccessExpiration = 'SetAccessExpiration',
  RegulateAsset = 'RegulateAsset',
  RegisterAndRegulateAsset = 'RegisterAndRegulateAsset',
  BindRegulatedAsset = 'BindRegulatedAsset',
  IssueSoulBoundToken = 'IssueSoulBoundToken',
}